requirements.txt
# Install: python3 -m pip install -r requirements.txt
# If you see "externally-managed-environment" (PEP 668), use a venv
#   python3 -m venv .venv && .venv/bin/pip install -r requirements.txt
# or: python3 -m pip install --break-system-packages -r requirements.txt
python-dotenv>=1.0.0
apify-client>=1.6.0
